third_party/blink/web_tests/platform/generic/http/tests/security/popup-allowed-by-sandbox-is-sandboxed-control-expected.txt

CONSOLE MESSAGE: /PASS/
To run this test outside of DumpRenderTree, please disable your popup blocker!

If you change this test, please be sure to change popup-allowed-by-sandbox-is-sandboxed.html as well!
